API refactoring
* Client and DSL classes moved to Occi::API module
* Client files moved to lib/occi/api
