+++ lib/occi/api/client/helper.rb
