+++ lib/occi/api/client/templates/compute.erb
