+++ lib/occi/api/client/templates/network.erb
