+++ lib/occi/api/client/templates/os_tpl.erb
