+++ lib/occi/api/client/templates/resource_tpl.erb
