+++ lib/occi/api/client/templates/storage.erb
